# docs/openapi.yaml
openapi: 3.0.0
info:
  title: superglue AI API
  version: 1.0.0
  description: API for running superglue AI tools
  contact:
    name: superglue AI Support
    email: stefan@superglue.ai
servers:
  - url: https://api.superglue.ai/v1
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /runs/{runId}:
    get:
      summary: Get run status
      operationId: getRun
      tags:
        - Runs
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
          example: 7f3e9c1a-2b4d-4e8f-9a3b-1c5d7e9f2a4b
      responses:
        "200":
          description: Run details
          headers:
            X-Trace-Id:
              schema:
                type: string
              description: Request trace ID for debugging and log correlation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "404":
          description: Run not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /runs/{runId}/cancel:
    post:
      summary: Cancel a run
      operationId: cancelRun
      tags:
        - Runs
      parameters:
        - name: runId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run cancelled
          headers:
            X-Trace-Id:
              schema:
                type: string
              description: Request trace ID for debugging and log correlation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Cannot cancel run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /runs:
    get:
      summary: List runs
      operationId: listRuns
      tags:
        - Runs
      parameters:
        - name: toolId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [running, success, failed, aborted]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        "200":
          description: List of runs
          headers:
            X-Trace-Id:
              schema:
                type: string
              description: Request trace ID for debugging and log correlation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Run"
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 50
                  total:
                    type: integer
                    example: 327
                  hasMore:
                    type: boolean
                    example: true
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: Static API Key
      description: |
        Static API key authentication using Bearer token scheme.
        Include your API key in the Authorization header: `Authorization: Bearer YOUR_API_KEY`

        Alternatively, you can use the `token` query parameter to authenticate.

        API keys can be generated in your superglue dashboard.
  schemas:
...
    Run:
      type: object
      required:
        - runId
        - toolId
        - status
        - metadata
      properties:
        runId:
          type: string
          description: Unique identifier for this run (maps to 'id' in GraphQL schema)
          example: 7f3e9c1a-2b4d-4e8f-9a3b-1c5d7e9f2a4b
        toolId:
          type: string
          description: ID of the tool that was executed
          example: 550e8400-e29b-41d4-a716-446655440000
        tool:
          type: object
          description: Tool metadata (not full configuration, maps to 'toolConfig' in GraphQL)
          required:
            - id
          properties:
            id:
              type: string
            version:
              type: string
              example: "2.1.0"
        status:
          type: string
          enum: [running, success, failed, aborted]
          description: |
            Execution status:
            - running: Execution in progress
            - success: Completed successfully
            - failed: Failed due to error
            - aborted: Cancelled by user or system
          example: success
        toolPayload:
          type: object
          description: The inputs and options provided when running the tool (maps to 'toolPayload' in GraphQL)
          additionalProperties: true
        data:
          type: object
          description: Tool execution results (only present when status is success, maps to 'toolResult' in GraphQL)
          additionalProperties: true
        error:
          type: string
          description: Error message (only present when status is failed or aborted)
          example: Connection timeout after 30000 milliseconds
        stepResults:
          type: array
          description: Results from each execution step (only for multi-step tools)
          items:
            type: object
            required:
              - stepId
              - success
            properties:
              stepId:
                type: string
                description: ID of the step that was executed
              success:
                type: boolean
                description: Whether the step completed successfully
              data:
                type: object
                description: Step execution result data
                additionalProperties: true
              error:
                type: string
                description: Error message if step failed
        options:
          type: object
          description: Execution options that were used for this run
          additionalProperties: true
        requestSource:
          type: string
          description: Source identifier for where the run was initiated
          example: api
        traceId:
          type: string
          description: Trace ID for this run (for debugging and log correlation)
          example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
        metadata:
          type: object
          properties:
            startedAt:
              type: string
              format: date-time
            completedAt:
              type: string
              format: date-time
              description: Only present when run has finished (success, failed, or aborted)
            durationMs:
              type: integer
              example: 5234
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
              example: Invalid input parameters
...